requirements.txt
asammdf
attrs
canmatrix
click
colorama
isal
lxml
lz4
numexpr
numpy
pandas
python-dateutil
pytz
six
typing_extensions
tzdata
matplotlib
fpdf2
scipy